requirements.txt
Pillow
google-generativeai
python-dotenv
asyncio
websockets
json